requirements.txt
streamlit
pandas
numpy
bcrypt
psycopg2-binary